pipeline/requirements.txt
pylint
pytest
python-dotenv
psycopg2
